Add integration tests for AWSMachine controller
